requirements.txt
django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
django-environ==0.8.1
pillow==8.4.0